fix: handle some webtiles race conditions
Most of this was found while debugging cao behavior when out of disk
space. A number of extremely strange race conditions could happen when
game startup fails in unusual ways, as well as in a subsequent attempt
to shutdown the server. This commit tries to handle these a bit better.
Unfortunately, the best solution I found for the big one (a race
condition after fork, but before execvpe overwrites python) is simply to
delay so that execvpe takes hold; fork + asyncio is not well behaved.
See:
    https://github.com/python/cpython/issues/66197
As far as I can tell, the promised solutions of safely closing the event
loop in a child never materialized.

+++ crawl-ref/source/webserver/webtiles/process_handler.py
@@ -605,18 +605,32 @@ class CrawlProcessHandlerBase(object):
         game_html = to_unicode(templ.generate(version = v))
         watcher.send_message("game_client", version = v, content = game_html)
 
-    def stop(self):
+    def stop(self, delay=False):
+        if delay:
+            IOLoop.current().call_later(0.2, self.stop)
+            return
         if self.process:
             self.process.flush_ttyrec()
-            self.process.send_signal(subprocess.signal.SIGHUP)
+            try:
+                self.process.send_signal(subprocess.signal.SIGHUP)
+            except OSError as e:
+                self.logger.error(f"Error {repr(e)} on SIGHUP to child process")
+                self._on_process_end()
+                return
             t = time.time() + config.get('kill_timeout')
             self.kill_timeout = IOLoop.current().add_timeout(t, self.kill)
 
     def kill(self):
         if self.process:
             self.logger.info("Killing crawl process after SIGHUP did nothing.")
-            self.process.send_signal(subprocess.signal.SIGABRT)
-            self.kill_timeout = None
+            try:
+                self.process.send_signal(subprocess.signal.SIGABRT)
+            except OSError as e:
+                self.logger.error(f"Error {repr(e)} on SIGKILL to child process")
+                self._on_process_end()
+                return
+            finally:
+                self.kill_timeout = None
 
     interesting_info = ("xl", "char", "place", "turn", "dur", "god", "title")
 
@@ -745,9 +759,9 @@ class CrawlProcessHandler(CrawlProcessHandlerBase):
     def start(self):
         self._purge_locks_and_start(True)
 
-    def stop(self):
+    def stop(self, delay=False):
         # n.b. the super call here is partly async
-        super(CrawlProcessHandler, self).stop()
+        super(CrawlProcessHandler, self).stop(delay=delay)
         self._stop_purging_stale_processes()
         self._stale_pid = None
 
@@ -916,6 +930,8 @@ class CrawlProcessHandler(CrawlProcessHandlerBase):
         else:
             self.logger.info("Starting game.")
 
+        connected = False
+
         try:
             self.process = TerminalRecorder(call,
                                             self.logger,
@@ -947,7 +963,12 @@ class CrawlProcessHandler(CrawlProcessHandlerBase):
             self.exit_dump_url = None
 
             if self.process and self.process.is_started():
-                self.stop()
+                # n.b. we delay a bit here so that the process has more
+                # time to start up, and avoid race conditions. (I couldn't come
+                # up with anything more reliable.) Also, currently
+                # if the connection fails the crawl process will be in a state
+                # where it ignores HUP, and the kill handler is needed.
+                self.stop(delay=True)
             else:
                 self._on_process_end()
